Include pointer to segment failing in error
On larger EDIFACT interchanges it will be a big help with a reference to the
segment number failing i.e. segment 42. Notice that the pointer (:segment)
starts from the end of the list of segments due to technical reasons.

During schemafication already processed segments are discarded thus it is
impossible to know how many already got processed.